Created preprocessor for FLAGS.h in order to setup hashes
Committing change before setting up the main overwrite.

Was creating this worth it? I don't know. It was interesting though

When adding new FLAGS just add to ATOM__FLAGS enum in FLAGS.h, run the preprocessor and the file should be updated with the hash and switch case. If anything goes wrong then revert in commit I guess. It should hopefully only overwrite if everything has worked with the new file though.